#89: build: rename examples
